requirements.txt
flask
marshmallow
flask_sqlalchemy
flask_smorest
python-dotenv
flask-migrate
